+++ R/utilities.R
@@ -394,7 +394,8 @@ check_args <- function(){
 #'   \code{model}: must be a length-1 \code{character} vector in
 #'     \code{\link{plot_cast_point}}, \code{\link{plot_cast_ts}},
 #'     \code{\link{plot_cast_ts_ylab}},
-#'     \code{\link{select_most_ab_spp}} \cr \cr
+#'     \code{\link{select_most_ab_spp}},
+#'     \code{\link{model_options}}, \code{\link{save_forecast_output}} \cr \cr
 #'   \code{models}: must be \code{NULL} or a \code{character} vector in
 #'     \code{\link{all_options}}, \code{\link{models_options}},
 #'     \code{\link{cast_options}}, \code{\link{model_paths}},
@@ -413,8 +414,6 @@ check_args <- function(){
 #'     \cr \cr
 #'   \code{nadot}: must be a length-1 \code{logical} vector in
 #'     \code{\link{rodent_spp}} \cr \cr
-#'   \code{name}: must be a length-1 \code{character} vector in
-#'     \code{\link{model_options}}, \code{\link{save_forecast_output}} \cr \cr
 #'   \code{ndates}: must be a length-1 positive \code{integer} or 
 #'     \code{integer}-conformable vector in 
 #'     \code{\link{plot_err_lead_spp_mods}}  \cr \cr
@@ -1138,14 +1137,6 @@ check_arg <- function(arg_name, arg_value, fun_name = NULL){
     }
     if (length(arg_value) != 1){
       stop("`nadot` can only be of length = 1")
-    }
-  }
-  if (arg_name == "name"){
-    if (!("character" %in% class(arg_value))){
-      stop("`name` is not a character")
-    }
-    if (length(arg_value) != 1){
-      stop("`name` can only be of length = 1")
     }
   }
   if (arg_name == "ndates"){